fixed payroll url

+++ app/Http/Controllers/PayrollController.php
@@ -6,7 +6,6 @@ use App\Http\Requests\PayrollRequest;
 use App\Http\Resources\BaseResource;
 use App\Models\Company;
 use App\Models\Payroll;
-use App\Models\Period;
 use App\Services\PayrollService;
 use App\Traits\Filter;
 use Illuminate\Http\JsonResponse;
@@ -25,12 +24,27 @@ class PayrollController extends Controller
 
     public function index(Request $request, Company $company): JsonResponse
     {
-        $payrolls = $this->applyFilters($request, $company->payrolls()->with('employee'), [
+        $payrollQuery = $company->payrolls();
+        if ($request->has('employee_id')) {
+            $payrollQuery->where('employee_id', $request->employee_id);
+        }
+        if ($request->has('period_id')) {
+            $payrollQuery->where('period_id', $request->period_id);
+        }
+        $payrolls = $this->applyFilters($request, $payrollQuery->with('employee'), [
             'status',
             'employee.first_name',
             'employee.last_name'
         ]);
         return $this->sendResponse($payrolls, 'Payrolls retrieved successfully.');
+    }
+
+    public function show(Company $company, Payroll $payroll): JsonResponse
+    {
+        if (!$company->payrolls->contains($payroll)) {
+            return $this->sendError('Payroll not found.');
+        }
+        return $this->sendResponse(new BaseResource($payroll), 'Payroll retrieved successfully.');
     }
     
     public function store(PayrollRequest $request, Company $company, int $employeeId): JsonResponse
@@ -39,14 +53,7 @@ class PayrollController extends Controller
         $employee = $company->getEmployeeById($employeeId);
         $period = $company->period($request->period_id);
         $payroll = $this->payrollService->generate($period, $employee, $input);
-        return $this->sendResponse(new BaseResource($payroll), 'Payroll created successfully');
-    }
-
-    public function show(Company $company, int $employeeId, int $payrollId): JsonResponse
-    {
-        $employee = $company->getEmployeeById($employeeId);
-        $payroll = $employee->payrolls()->where('id', $payrollId)->first();
-        return $this->sendResponse(new BaseResource($payroll), 'Payroll retrieved successfully');
+        return $this->sendResponse(new BaseResource($payroll), 'Payroll created successfully.');
     }
 
     public function update(PayrollRequest $request, Company $company, int $employeeId, int $payrollId): JsonResponse
@@ -60,11 +67,12 @@ class PayrollController extends Controller
         return $this->sendResponse(new BaseResource($payroll), 'Payroll updated successfully.');
     }
 
-    public function destroy(Company $company, int $employeeId, int $payrollId): JsonResponse
+    public function destroy(Company $company, Payroll $payroll): JsonResponse
     {
-        $employee = $company->getEmployeeById($employeeId);
-        $payroll = $employee->payrolls()->where('id', $payrollId)->first();
+        if (!$company->payrolls->contains($payroll)) {
+            return $this->sendError('Payroll not found.');
+        }
         $payroll->delete();
-        return $this->sendResponse(new BaseResource($payroll), 'Payroll deleted successfully');
+        return $this->sendResponse(new BaseResource($payroll), 'Payroll deleted successfully.');
     }
 }
